pyproject.toml
[project]
name = "code-analyser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gcloud>=0.18.3",
    "google-adk>=1.18.0",
    "google-genai>=1.50.1",
    "langchain>=1.0.7",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
]
